[#6] Chore: modal 예시 test페이지로 옮김
